fix: refine duckdb testing (#1019)
minor refactorings for duckdb testing
